[REFACTOR] Дублирующийся код #9
